m1 project home page requirements
A link to the GitHub organization associated with this project and all of its repositories.

A section called “Deployment” containing a link to the deployed application running on Digital Ocean.

Up-to-date screen shots showing the state of the project.

A link to the M1 Project page, showing what issues were completed for this milestone. It is appropriate that by the time of the due date, there are no issues in the BackLog or In Progress for this milestone.

A link to the M2 Project page, showing the issues expected to be addressed during the second Milestone for this project.

Adherence to the GitHub hosting guidelines.